requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
argon2-cffi==23.1.0
python-dotenv==1.0.0
requests==2.31.0
captcha==0.4
Pillow==9.5.0
